# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/specification/test": {
            "get": {
                "description": "Check if domain account can be verified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specification"
                ],
                "summary": "Test auth against the specification",
                "parameters": [
                    {
                        "description": "Enter auth and domain name properties",
                        "name": "text",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.Specification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Specification": {
            "type": "object",
            "properties": {
                "domain_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}